Algorithms/src/main/java/szczesnaj/InsertionSort.java
/*
 * Copyright (c) 2023 Joanna Szczesna
 * All rights reserved
 */

package szczesnaj;

class InsertionSort {

    static int[] sort(int[] arr) {
        if (arr == null) {
            return null;
        }
        for (int i = 1; i <= arr.length - 1; i++) {
            for (int j = i; j >= 1; j--) {
                if (arr[j] < arr[j - 1]) {
                    int temp = arr[j];
                    arr[j] = arr[j - 1];
                    arr[j - 1] = temp;
                }
            }
        }
        return arr;
    }
}
